+++ src/app/src/test/kotlin/org/imperial/mrc/hint/unit/security/SessionTests.kt
@@ -1,8 +1,6 @@
 package org.imperial.mrc.hint.unit.security
 
-import com.nhaarman.mockito_kotlin.doReturn
-import com.nhaarman.mockito_kotlin.mock
-import com.nhaarman.mockito_kotlin.verify
+import com.nhaarman.mockito_kotlin.*
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.pac4j.core.config.Config
@@ -47,7 +45,7 @@ class SessionTests
         val mockSessionStore = mock<SessionStore<WebContext>>()
         val mockConfig = mock<Config>
         {
-            on {sessionStore} doReturn mockSessionStore
+            on { sessionStore } doReturn mockSessionStore
         }
         val sut = Session(mockWebContext, mockConfig)
         val versionId = sut.getVersionId()
@@ -55,5 +53,73 @@ class SessionTests
 
         verify(mockSessionStore).set(mockWebContext, "version_id", versionId)
     }
-}
 
+    @Test
+    fun `setMode saves new mode and clears version id if mode has changed`()
+    {
+        val mockWebContext = mock<WebContext>()
+        val mockSessionStore = mock<SessionStore<WebContext>> {
+            on { get(mockWebContext, "mode") } doReturn "naomi"
+        }
+        val mockConfig = mock<Config>
+        {
+            on { sessionStore } doReturn mockSessionStore
+        }
+
+        val sut = Session(mockWebContext, mockConfig)
+        sut.setMode("explore")
+
+        verify(mockSessionStore).set(mockWebContext, "mode", "explore")
+        verify(mockSessionStore).set(mockWebContext, "version_id", null)
+    }
+
+    @Test
+    fun `setMode does nothing if mode has not changed`()
+    {
+        val mockWebContext = mock<WebContext>()
+        val mockSessionStore = mock<SessionStore<WebContext>> {
+            on { get(mockWebContext, "mode") } doReturn "naomi"
+        }
+        val mockConfig = mock<Config>
+        {
+            on { sessionStore } doReturn mockSessionStore
+        }
+
+        val sut = Session(mockWebContext, mockConfig)
+        sut.setMode("naomi")
+
+        verify(mockSessionStore, never()).set(any(), any(), any())
+    }
+
+    @Test
+    fun `can set pac4j requested url to string value`()
+    {
+        val mockWebContext = mock<WebContext>()
+        val mockSessionStore = mock<SessionStore<WebContext>>()
+        val mockConfig = mock<Config>
+        {
+            on { sessionStore } doReturn mockSessionStore
+        }
+
+        val sut = Session(mockWebContext, mockConfig)
+        sut.setRequestedUrl("explore")
+
+        verify(mockSessionStore).set(mockWebContext, "pac4jRequestedUrl", "explore")
+    }
+
+    @Test
+    fun `can set pac4j requested url to null`()
+    {
+        val mockWebContext = mock<WebContext>()
+        val mockSessionStore = mock<SessionStore<WebContext>>()
+        val mockConfig = mock<Config>
+        {
+            on { sessionStore } doReturn mockSessionStore
+        }
+
+        val sut = Session(mockWebContext, mockConfig)
+        sut.setRequestedUrl(null)
+
+        verify(mockSessionStore).set(mockWebContext, "pac4jRequestedUrl", null)
+    }
+}