+++ src/app/static/src/app/store/language/mutations.ts
@@ -1,15 +1,19 @@
 import {MutationTree} from "vuex";
-import {PayloadWithType} from "../../types";
+import {PayloadWithType, TranslatableState} from "../../types";
 import {Language} from "../translations/locales";
-import {TranslatableState} from "../../root";
 
 export enum LanguageMutation {
-    ChangeLanguage = "ChangeLanguage"
+    ChangeLanguage = "ChangeLanguage",
+    SetUpdatingLanguage = "SetUpdatingLanguage"
 }
 
 export const mutations: MutationTree<TranslatableState> = {
 
     [LanguageMutation.ChangeLanguage](state: TranslatableState, action: PayloadWithType<Language>) {
         state.language = action.payload
+    },
+
+    [LanguageMutation.SetUpdatingLanguage](state: TranslatableState, action: PayloadWithType<boolean>) {
+        state.updatingLanguage = action.payload;
     }
 };