+++ src/app/static/src/app/store/plottingSelections/getters.ts
@@ -1,9 +1,9 @@
 import {DataType} from "../surveyAndProgram/surveyAndProgram";
-import {RootState} from "../../root";
 import {ScaleSelections, PlottingSelectionsState, BarchartSelections} from "./plottingSelections";
+import {DataExplorationState} from "../dataExploration/dataExploration";
 
 export const getters = {
-    selectedSAPColourScales: (state: PlottingSelectionsState, getters: any, rootState: RootState): ScaleSelections => {
+    selectedSAPColourScales: (state: PlottingSelectionsState, getters: any, rootState: DataExplorationState): ScaleSelections => {
         const dataType = rootState.surveyAndProgram.selectedDataType;
         switch (dataType) {
             case DataType.Survey:
@@ -15,8 +15,5 @@ export const getters = {
             default:
                 return {}
         }
-    },
-    calibratePlotDefaultSelections: (state: PlottingSelectionsState, getters: any, rootState: RootState): BarchartSelections => {
-        return rootState.modelCalibrate.calibratePlotResult!.plottingMetadata.barchart.defaults;
     }
 };