+++ src/app/static/src/app/store/plottingSelections/plottingSelections.ts
@@ -1,10 +1,10 @@
 import {FilterOption} from "../../generated";
 import {localStorageManager} from "../../localStorageManager";
 import {Module} from "vuex";
-import {RootState} from "../../root";
 import {mutations} from "./mutations";
 import {getters} from "./getters";
 import {Dict} from "../../types";
+import {DataExplorationState} from "../dataExploration/dataExploration";
 
 export interface PlottingSelectionsState {
     calibratePlot: BarchartSelections,
@@ -140,7 +140,7 @@ export const initialPlottingSelectionsState = (): PlottingSelectionsState => {
 const namespaced = true;
 const existingState = localStorageManager.getState();
 
-export const plottingSelections: Module<PlottingSelectionsState, RootState> = {
+export const plottingSelections: Module<PlottingSelectionsState, DataExplorationState> = {
     namespaced,
     state: {...initialPlottingSelectionsState(), ...existingState && existingState.plottingSelections},
     mutations,