+++ src/app/static/src/app/store/surveyAndProgram/actions.ts
@@ -1,26 +1,32 @@
 import {ActionContext, ActionTree, Commit} from 'vuex';
-import {RootState} from "../../root";
 import {DataType, SurveyAndProgramState} from "./surveyAndProgram";
 import {api} from "../../apiService";
 import {AncResponse, ProgrammeResponse, SurveyResponse} from "../../generated";
 import {SurveyAndProgramMutation} from "./mutations";
 import qs from 'qs';
 import {getFilenameFromImportUrl, getFilenameFromUploadFormData} from "../../utils";
+import {GenericChartMutation} from "../genericChart/mutations";
+import {DataExplorationState} from "../dataExploration/dataExploration";
 
 export interface SurveyAndProgramActions {
-    importSurvey: (store: ActionContext<SurveyAndProgramState, RootState>, url: string) => void,
-    importProgram: (store: ActionContext<SurveyAndProgramState, RootState>, url: string) => void,
-    importANC: (store: ActionContext<SurveyAndProgramState, RootState>, url: string) => void,
-    uploadSurvey: (store: ActionContext<SurveyAndProgramState, RootState>, formData: FormData) => void,
-    uploadProgram: (store: ActionContext<SurveyAndProgramState, RootState>, formData: FormData) => void,
-    uploadANC: (store: ActionContext<SurveyAndProgramState, RootState>, formData: FormData) => void
-    getSurveyAndProgramData: (store: ActionContext<SurveyAndProgramState, RootState>) => void;
-    deleteSurvey: (store: ActionContext<SurveyAndProgramState, RootState>) => void
-    deleteProgram: (store: ActionContext<SurveyAndProgramState, RootState>) => void
-    deleteANC: (store: ActionContext<SurveyAndProgramState, RootState>) => void
-    deleteAll: (store: ActionContext<SurveyAndProgramState, RootState>) => void
-    selectDataType: (store: ActionContext<SurveyAndProgramState, RootState>, payload: DataType) => void
-    validateSurveyAndProgramData: (store: ActionContext<SurveyAndProgramState, RootState>) => void;
+    importSurvey: (store: ActionContext<SurveyAndProgramState, DataExplorationState>, url: string) => void,
+    importProgram: (store: ActionContext<SurveyAndProgramState, DataExplorationState>, url: string) => void,
+    importANC: (store: ActionContext<SurveyAndProgramState, DataExplorationState>, url: string) => void,
+    uploadSurvey: (store: ActionContext<SurveyAndProgramState, DataExplorationState>, formData: FormData) => void,
+    uploadProgram: (store: ActionContext<SurveyAndProgramState, DataExplorationState>, formData: FormData) => void,
+    uploadANC: (store: ActionContext<SurveyAndProgramState, DataExplorationState>, formData: FormData) => void
+    getSurveyAndProgramData: (store: ActionContext<SurveyAndProgramState, DataExplorationState>) => void;
+    deleteSurvey: (store: ActionContext<SurveyAndProgramState, DataExplorationState>) => void
+    deleteProgram: (store: ActionContext<SurveyAndProgramState, DataExplorationState>) => void
+    deleteANC: (store: ActionContext<SurveyAndProgramState, DataExplorationState>) => void
+    deleteAll: (store: ActionContext<SurveyAndProgramState, DataExplorationState>) => void
+    selectDataType: (store: ActionContext<SurveyAndProgramState, DataExplorationState>, payload: DataType) => void
+    validateSurveyAndProgramData: (store: ActionContext<SurveyAndProgramState, DataExplorationState>) => void;
+}
+
+const enum DATASET_TYPE  {
+    ANC = "anc",
+    ART = "art"
 }
 
 function commitSelectedDataTypeUpdated(commit: Commit, dataType: DataType) {
@@ -28,15 +34,20 @@ function commitSelectedDataTypeUpdated(commit: Commit, dataType: DataType) {
         {type: "SelectedDataTypeUpdated", payload: dataType}, {root: true})
 }
 
+function commitClearGenericChartDataset(commit: Commit, dataType: string) {
+    commit({type: `genericChart/${GenericChartMutation.ClearDataset}`, payload: dataType}, {root: true});
+}
+
 interface UploadImportOptions {
     url: string
     payload: FormData | string
 }
 
-async function uploadOrImportANC(context: ActionContext<SurveyAndProgramState, RootState>, options: UploadImportOptions,
+async function uploadOrImportANC(context: ActionContext<SurveyAndProgramState, DataExplorationState>, options: UploadImportOptions,
                                  filename: string) {
     const {commit} = context;
     commit({type: SurveyAndProgramMutation.ANCUpdated, payload: null});
+    commitClearGenericChartDataset(commit, DATASET_TYPE.ANC);
 
     await api<SurveyAndProgramMutation, SurveyAndProgramMutation>(context)
         .withError(SurveyAndProgramMutation.ANCError)
@@ -52,10 +63,11 @@ async function uploadOrImportANC(context: ActionContext<SurveyAndProgramState, R
         });
 }
 
-async function uploadOrImportProgram(context: ActionContext<SurveyAndProgramState, RootState>, options: UploadImportOptions,
+async function uploadOrImportProgram(context: ActionContext<SurveyAndProgramState, DataExplorationState>, options: UploadImportOptions,
                                      filename: string) {
     const {commit} = context;
     commit({type: SurveyAndProgramMutation.ProgramUpdated, payload: null});
+    commitClearGenericChartDataset(commit, DATASET_TYPE.ART);
 
     await api<SurveyAndProgramMutation, SurveyAndProgramMutation>(context)
         .withError(SurveyAndProgramMutation.ProgramError)
@@ -71,7 +83,7 @@ async function uploadOrImportProgram(context: ActionContext<SurveyAndProgramStat
         });
 }
 
-async function uploadOrImportSurvey(context: ActionContext<SurveyAndProgramState, RootState>, options: UploadImportOptions,
+async function uploadOrImportSurvey(context: ActionContext<SurveyAndProgramState, DataExplorationState>, options: UploadImportOptions,
                                     filename: string) {
     const {commit} = context;
     commit({type: SurveyAndProgramMutation.SurveyUpdated, payload: null});
@@ -90,7 +102,7 @@ async function uploadOrImportSurvey(context: ActionContext<SurveyAndProgramState
         });
 }
 
-export const actions: ActionTree<SurveyAndProgramState, RootState> & SurveyAndProgramActions = {
+export const actions: ActionTree<SurveyAndProgramState, DataExplorationState> & SurveyAndProgramActions = {
 
     selectDataType(context, payload) {
         const {commit} = context;
@@ -142,6 +154,7 @@ export const actions: ActionTree<SurveyAndProgramState, RootState> & SurveyAndPr
             .delete("/disease/programme/")
             .then(() => {
                 commit({type: SurveyAndProgramMutation.ProgramUpdated, payload: null});
+                commitClearGenericChartDataset(commit, DATASET_TYPE.ART)
             });
     },
 
@@ -151,6 +164,7 @@ export const actions: ActionTree<SurveyAndProgramState, RootState> & SurveyAndPr
             .delete("/disease/anc/")
             .then(() => {
                 commit({type: SurveyAndProgramMutation.ANCUpdated, payload: null});
+                commitClearGenericChartDataset(commit, DATASET_TYPE.ANC)
             });
     },